requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
WTForms==3.1.1
python-dotenv==1.0.0
numpy>=1.25.0
pandas==2.3.3
openpyxl==3.1.2
gunicorn==20.1.0
serverless-wsgi==3.1.0
Werkzeug>=3.0.0
